feat(backup): Separate uploadMedia platform specific methods into two files
By leveraging platform specific extensions, we will allow the bundler to remove unused methods and as a result unused RN modules used to upload medias.

I needed to update tsconfig.json to support these extensions.
