+++ tsconfig.json
@@ -7,5 +7,6 @@
     "paths": {
       "/*": ["src/*"]
     },
+    "moduleSuffixes": [".ios", ".android", ""]
   }
 }
